backend/grpc-server/requirements.txt
grpcio==1.32.0
grpcio-tools==1.32.0
pydevd-pycharm~=201.7846.77
pytz==2020.1